PythonQAOtus_Lesson19/Study/requirements.txt
attrs==19.3.0
Cerberus==1.3.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Eve==1.1.1
Events==0.3
Flask==1.1.2
httpie==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
mysql-connector==2.2.9
mysql-connector-python==8.0.20
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
psycopg2-binary==2.8.5
py==1.8.1
Pygments==2.6.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
requests==2.23.0
simplejson==3.17.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
